moving stateproofverification tracker's sql function into a file.
